refactor(backend): extract backend-mode state into a leaf module
Move configuredBackendMode + resolveBackendMode out of the 540-line backend.ts
into a focused backend-mode.ts leaf, so the active-mode resolver no longer sits
among the backend-class imports. backend.ts re-exports BackendMode and
isExperimentalLocalBackendEnabled, so consumers via the @/backend barrel are
unaffected.

configureBackendMode keeps the resolved mode and the
LETTA_LOCAL_BACKEND_EXPERIMENTAL env in sync, so the LRU/session getters already
namespace by the active server key — the startup resume path's extra
isAgentIdCompatibleWithBackend re-checks were dead. Remove all five plus the now
-unused import; the mode guards alone suffice.

Settings namespacing intentionally stays on the env-based predicate
(isLocalBackendEnvEnabled): keying it on the mutable resolved mode would couple
settings to global state that leaks across test files (verified via a startup
fallback test that only fails in the shared-worker suite).
